requirements.txt
pyTelegramBotAPI==4.14.0
mysql-connector-python==8.2.0 